requirements.txt
pandas
pyyaml
pytorch-lightning
transformers
transformers[torch]
rouge
rouge-score
wandb
tqdm
scikit-learn
openai
Faker
datasets